requirements.txt
numpy==1.24.3
tensorflow==2.10
keras==2.10
Pillow==10.0.1
scikit-learn==1.3.1
opencv-python>4.7.0
flowiz==2.4.0